# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase GraphQL API",
    "description": "API for managing companies, establishments, officers, addresses, files, and related data using GraphQL with a Supabase backend. All operations are performed via a single GraphQL endpoint using POST requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local GraphQL server"
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": { "summary": "Get all users", "tags": ["User"], "responses": { "200": { "description": "List of users", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } } },
      "post": { "summary": "Create a new user", "tags": ["User"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } } } }, "responses": { "201": { "description": "User created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/users/{id}": {
      "get": { "summary": "Get a user by ID", "tags": ["User"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "User found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "description": "User not found" } } },
      "put": { "summary": "Update a user by ID", "tags": ["User"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserInput" } } } }, "responses": { "200": { "description": "User updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "User not found" } } },
      "delete": { "summary": "Delete a user by ID", "tags": ["User"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "User deleted" }, "404": { "description": "User not found" } } }
    },
...
    "/api/risk_assessments": {
      "get": { "summary": "Get all risk assessments", "tags": ["RiskAssessment"], "responses": { "200": { "description": "List of risk assessments", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/RiskAssessment" } } } } } } },
      "post": { "summary": "Create a new risk assessment", "tags": ["RiskAssessment"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskAssessmentInput" } } } }, "responses": { "201": { "description": "Risk assessment created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskAssessment" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/risk_assessments/{id}": {
      "get": { "summary": "Get a risk assessment by ID", "tags": ["RiskAssessment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Risk assessment found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskAssessment" } } } }, "404": { "description": "Risk assessment not found" } } },
      "put": { "summary": "Update a risk assessment by ID", "tags": ["RiskAssessment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskAssessmentInput" } } } }, "responses": { "200": { "description": "Risk assessment updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskAssessment" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Risk assessment not found" } } },
      "delete": { "summary": "Delete a risk assessment by ID", "tags": ["RiskAssessment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Risk assessment deleted" }, "404": { "description": "Risk assessment not found" } } }
    },
...
    "/api/subscriptions": {
      "get": { "summary": "Get all subscriptions", "tags": ["Subscription"], "responses": { "200": { "description": "List of subscriptions", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Subscription" } } } } } } },
      "post": { "summary": "Create a new subscription", "tags": ["Subscription"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubscriptionInput" } } } }, "responses": { "201": { "description": "Subscription created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/subscriptions/{id}": {
      "get": { "summary": "Get a subscription by ID", "tags": ["Subscription"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Subscription found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } }, "404": { "description": "Subscription not found" } } },
      "put": { "summary": "Update a subscription by ID", "tags": ["Subscription"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubscriptionInput" } } } }, "responses": { "200": { "description": "Subscription updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Subscription not found" } } },
      "delete": { "summary": "Delete a subscription by ID", "tags": ["Subscription"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Subscription deleted" }, "404": { "description": "Subscription not found" } } }
    },
...
  },
  "components": {
    "schemas": {
      "Company": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "registration_number": { "type": "string", "description": "Company registration number (max 20 characters)" },
          "full_registration_number": { "type": "string", "description": "Full registration number including country code (max 20 characters)" },
          "vat_number": { "type": "string", "nullable": true, "description": "VAT number (max 50 characters)" },
          "name": { "type": "string", "description": "Company name" },
          "business_name": { "type": "string", "nullable": true, "description": "Business or trading name" },
          "creation_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of company creation" },
          "status": { "type": "string", "nullable": true, "description": "Company status (e.g., active, dissolved)" },
          "capital": { "type": "number", "nullable": true, "description": "Company capital amount" },
          "establishments_count": { "type": "integer", "nullable": true, "description": "Number of establishments" },
          "last_legal_update": { "type": "string", "format": "date", "nullable": true, "description": "Date of last legal update" },
          "establishments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Establishment" },
            "description": "List of company establishments"
          },
          "officers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Officer" },
            "description": "List of company officers"
          },
          "beneficial_owners": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BeneficialOwner" },
            "description": "List of beneficial owners"
          },
          "financial_statements": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FinancialStatement" },
            "description": "List of financial statements"
          },
          "risk_assessment": { "$ref": "#/components/schemas/RiskAssessment", "nullable": true, "description": "Risk assessment for the company" },
          "bodacc_notices": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/BodaccNotice" },
            "description": "List of BODACC notices"
          },
          "legal_acts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LegalAct" },
            "description": "List of legal acts"
          },
          "web_info": { "$ref": "#/components/schemas/WebInfo", "nullable": true, "description": "Web and social media information" },
          "emails": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Email" },
            "description": "List of associated email addresses"
          }
        },
        "required": ["id", "registration_number", "full_registration_number", "name"]
      },
...
      "Establishment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "siret": { "type": "string", "description": "SIRET number (max 20 characters)" },
          "is_head_office": { "type": "boolean", "description": "Indicates if this is the head office" },
          "creation_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of establishment creation" },
          "status": { "type": "string", "nullable": true, "description": "Establishment status" },
          "address_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Reference to address" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "address": { "$ref": "#/components/schemas/Address", "nullable": true, "description": "Associated address" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "siret", "is_head_office", "company_id"]
      },
...
      "Officer": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "first_name": { "type": "string", "nullable": true, "description": "Officer's first name" },
          "last_name": { "type": "string", "nullable": true, "description": "Officer's last name" },
          "role": { "type": "string", "nullable": true, "description": "Officer's role (e.g., CEO, CFO)" },
          "start_date": { "type": "string", "format": "date", "nullable": true, "description": "Start date of role" },
          "end_date": { "type": "string", "format": "date", "nullable": true, "description": "End date of role" },
          "registration_number": { "type": "string", "nullable": true, "description": "Registration number if officer is a company (max 20 characters)" },
          "is_company": { "type": "boolean", "nullable": true, "description": "Indicates if the officer is a company" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
...
      "BeneficialOwner": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "first_name": { "type": "string", "nullable": true, "description": "Beneficial owner's first name" },
          "last_name": { "type": "string", "nullable": true, "description": "Beneficial owner's last name" },
          "birth_date": { "type": "string", "nullable": true, "description": "Birth date as text" },
          "direct_shares_pct": { "type": "number", "nullable": true, "description": "Percentage of direct shares" },
          "indirect_shares_pct": { "type": "number", "nullable": true, "description": "Percentage of indirect shares" },
          "voting_rights_pct": { "type": "number", "nullable": true, "description": "Percentage of voting rights" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
...
      "FinancialStatement": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "year": { "type": "integer", "description": "Financial year" },
          "sales": { "type": "number", "nullable": true, "description": "Sales amount" },
          "profit": { "type": "number", "nullable": true, "description": "Profit amount" },
          "staff": { "type": "integer", "nullable": true, "description": "Number of staff" },
          "closing_date": { "type": "string", "format": "date", "nullable": true, "description": "Financial statement closing date" },
          "is_estimated": { "type": "boolean", "nullable": true, "description": "Indicates if data is estimated" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "year", "company_id"]
      },
...
      "RiskAssessment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "risk_level": { "type": "integer", "minimum": 0, "maximum": 5, "description": "Risk level (0-5)" },
          "description": { "type": "string", "nullable": true, "description": "Risk assessment description" },
          "collective_procedure": { "type": "boolean", "nullable": true, "description": "Indicates if under collective procedure" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company (unique)" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "risk_level", "company_id"]
      },
      "RiskAssessmentInput": {
        "type": "object",
        "properties": {
          "risk_level": { "type": "integer", "minimum": 0, "maximum": 5, "description": "Risk level (0-5)" },
          "description": { "type": "string", "nullable": true, "description": "Risk assessment description" },
          "collective_procedure": { "type": "boolean", "nullable": true, "description": "Indicates if under collective procedure" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" }
        },
        "required": ["risk_level", "company_id"]
      },
      "BodaccNotice": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "type": { "type": "string", "nullable": true, "description": "Type of BODACC notice" },
          "parution_date": { "type": "string", "format": "date", "nullable": true, "description": "Publication date" },
          "number": { "type": "string", "nullable": true, "description": "Notice number" },
          "description": { "type": "string", "nullable": true, "description": "Notice description" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
...
      "LegalAct": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "date": { "type": "string", "format": "date", "nullable": true, "description": "Date of legal act" },
          "type": { "type": "string", "nullable": true, "description": "Type of legal act" },
          "description": { "type": "string", "nullable": true, "description": "Legal act description" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
...
      "WebInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "logos": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Array of logo URLs" },
          "websites": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Array of website URLs" },
          "linkedin_url": { "type": "string", "nullable": true, "description": "LinkedIn URL" },
          "twitter_url": { "type": "string", "nullable": true, "description": "Twitter URL" },
          "facebook_url": { "type": "string", "nullable": true, "description": "Facebook URL" },
          "instagram_url": { "type": "string", "nullable": true, "description": "Instagram URL" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company (unique)" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
...
      "Email": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "value": { "type": "string", "description": "Email address" },
          "type": { "type": "string", "enum": ["generic", "personal"], "description": "Email type" },
          "test_result": { "type": "string", "nullable": true, "description": "Result of email validation test" },
          "test_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of email validation test" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "value", "type", "company_id"]
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "street": { "type": "string", "nullable": true, "description": "Street address" },
          "postal_code": { "type": "string", "nullable": true, "description": "Postal code (max 10 characters)" },
          "city": { "type": "string", "nullable": true, "description": "City" },
          "country_code": { "type": "string", "nullable": true, "description": "Country code (max 5 characters)" },
          "lat": { "type": "number", "nullable": true, "description": "Latitude" },
          "lng": { "type": "number", "nullable": true, "description": "Longitude" }
        },
        "required": ["id"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "name": { "type": "string", "description": "User's name" },
          "email": { "type": "string", "description": "User's email" },
          "role": { "type": "string", "description": "User's role" },
          "status": { "type": "string", "description": "User's status" },
          "created_at": { "type": "string", "format": "date-time", "description": "Date and time user was created" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Date and time user was last updated" }
        },
        "required": ["id", "name", "email", "role", "status"]
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "User's name" },
          "email": { "type": "string", "description": "User's email" },
          "role": { "type": "string", "description": "User's role" },
          "status": { "type": "string", "description": "User's status" }
        },
        "required": ["name", "email", "role", "status"]
      },
...
    }
  },
...
}